Cleaning up schedule page, also correctly loading observables on Safari and Firefox so schedule renders. Did not fix the bug where re-ordering items in schedule causes the schedule component to blank out.
